Footerのコンポーネント化
